pyproject.toml
[project]
name = "starclinch-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.40.24",
    "celery>=5.5.3",
    "cloudinary>=1.44.1",
    "django>=4.1",
    "django-celery-beat>=2.8.1",
    "django-cloudinary-storage>=0.3.0",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
]
